refactor(cli): split main.rs into focused bootstrap, dispatch, and CLI modules [OPE-412]
Decompose the 755-line monolithic main.rs into focused modules:
- cli.rs: Cli struct, Command enum, CompletionShell, and completion output
- bootstrap.rs: rustls init, env var setup, profile/team registration, tokio runtime
- dispatch.rs: async command dispatch to subcommand handlers
- tests.rs: all 30 parser/CLI tests

main.rs is now a 30-line slim entrypoint that wires the modules together.
All existing CLI behavior, flags, and command semantics are preserved.
